requirements.txt
flask
flask-limiter
authlib
anthropic
markdown
python-dotenv
stripe
werkzeug
resend
pytest
